Refactoring navigation and changing to lower version react-navigation/stack
