+++ packages/mobile/package.json
@@ -19,7 +19,7 @@
     "@react-navigation/material-bottom-tabs": "^5.1.7",
     "@react-navigation/material-top-tabs": "^5.1.7",
     "@react-navigation/native": "^5.1.4",
-    "@react-navigation/stack": "^5.7.0",
+    "@react-navigation/stack": "5.2.9",
     "axios": "^0.19.1",
     "formik": "^2.1.4",
     "react": "16.9.0",
